Added authorization to API

+++ backend/controllers/productController.js
@@ -1,13 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
-const Product = require('../model/productModel')
-
+const Product = require('../models/productModel')
+const User = require('../models/userModel')
 
 //@ desc   GET PRODUCTS 
 // @route  GET /API/PRODUCTS
 // @access Private
 const getProducts = asyncHandler(async (req, res) => {
-    const products = await Product.find()
+    const products = await Product.find({ user: req.user.id })
 
     res.status(200).json({ products })
 })
@@ -22,7 +22,8 @@ const postProduct = asyncHandler(async (req, res) => {
     }
 
     const product = await Product.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json({ product })
@@ -37,6 +38,19 @@ const updateProduct = asyncHandler(async (req, res) => {
     if(!product){
         res.status(400)
         throw new Error('Product not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // Check For User
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure the logged in user matches the product user
+    if(product.user.toString()!== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedProduct = await Product.findByIdAndUpdate(req.params.id, req.body, {
@@ -56,6 +70,19 @@ const deleteProduct = asyncHandler(async (req, res) => {
     if(!product){
         res.status(400)
         throw new Error('Product not found')
+    }
+    
+    const user = await User.findById(req.user.id)
+
+    // Check For User
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //Make sure the logged in user matches the product user
+    if(product.user.toString()!== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await product.deleteOne()
